Add a demo program for CPP_FFFE#12
Adds a demo program that simulates a smartwatch.
The smartwatch has a GPS module that can be queried for
latitude/longitude coordinates. Here the measurements are emulated
as a file containing time offsets and distance from origin - which
removes the need to perform complicated lat/lon to distance
calculations.

The smartwatch needs to calculate the average velocity of the user,
and based on this average velocity, classify the exercise as either a
run (when avg velocity >= 3.0) or a walk (otherwise).

The data indicates that the exercise should be classified as a run.
However, that does not happen - this is due to a NaN poisoning.
The audience will be asked to find / try to guesstimate the source of
such an error.

This example strives to be as simple and yet as realistic as possible.