Add the auto navigation script
The inter-slide navigation is currently realized by the .html file where
each slide requires the user to manually specify the slide ordering number,
and the two navigation buttons.
This is not optimal. In this commit I'm introducing a javascript file
that will in the future automatically inject the two navigation buttons,
as well as the slide number.
So far the script gets the handle of the relevant footer element.