Add functionality for inserting slide numbers
This commit implements a different approach; one in which
all the slides are processed once, at startup.
This will finally result in a slide definition that will
not require providing a footer at all.
For the time being, it modifies the slide number part
of the footer on each individual slide.